Rename AssetRandomizer class to AssetNameObfuscator
